scripts/release/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "release-tool"
version = "0.1.0"
description = "Multi-ecosystem release automation tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Emasoft", email = "713559+Emasoft@users.noreply.github.com" }
]
keywords = ["release", "automation", "cli", "npm", "pypi", "cargo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Software Distribution",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
    "semver>=3.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
release = "release.cli:app"

[project.urls]
Homepage = "https://github.com/Emasoft/SVG-BBOX"
Repository = "https://github.com/Emasoft/SVG-BBOX"
Issues = "https://github.com/Emasoft/SVG-BBOX/issues"

[tool.hatch.build.targets.wheel]
packages = ["release"]

[tool.hatch.build.targets.sdist]
include = ["release/**/*.py", "README.md"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "pydantic",
    "pydantic_settings",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
